+++ CHANGELOG.md
@@ -29,7 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `StatImputer` to `impute`
 - `CrossEntropy` to `optim`
 - `PAClassifier`, `PARegressor`, and `SoftmaxRegression` to `linear_model`
-- `check_estimator`, `SortedWindow`, and `Window` to `utils`
+- `check_estimator`, `Histogram`, `SortedWindow`, and `Window` to `utils`
 
 ### Removed